server/python-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.0.3          # <-- compatible with Python 3.8
numpy==1.24.4          # <-- pair with pandas 2.0.x on Py3.8
scikit-learn==1.3.2    # ENGLISH_STOP_WORDS; Py3.8 compatible
rank-bm25==0.2.2
faiss-cpu==1.7.4
sentence-transformers==2.6.1
transformers>=4.38.0,<5
torch                   # install the build that matches your CUDA/CPU